[JS] Rename posts.js.erb -> posts.js
There was no templating stuff used anyways
