+++ app/javascript/src/javascripts/mod_queue.js
@@ -1,5 +1,5 @@
 import Utility from './utility'
-import Post from './posts.js.erb'
+import Post from './posts'
 
 let ModQueue = {};
