+++ app/javascript/src/javascripts/post_sets.js
@@ -1,5 +1,5 @@
 import {SendQueue} from './send_queue'
-import Post from './posts.js.erb'
+import Post from './posts'
 import LS from './local_storage'
 
 let PostSet = {};
