+++ app/javascript/src/javascripts/tag_editor.vue
@@ -30,7 +30,7 @@
   import Vue from 'vue';
   import relatedTags from './uploader_related.vue';
   import tagPreview from './uploader_preview.vue';
-  import Post from './posts.js.erb';
+  import Post from './posts';
   import Autocomplete from "./autocomplete.js.erb";
   import Utility from "./utility.js";
