+++ app/javascript/src/javascripts/uploads.js
@@ -1,4 +1,4 @@
-import Post from './posts.js.erb'
+import Post from './posts'
 
 let Upload = {};
